requirements.txt

pyserial
pandas
matplotlib
